requirements.txt
# Requirements for the Python project
# Add your project dependencies here.
pytest
Flask==2.3.3
Flask-SocketIO==5.3.4
ollama-python==0.1.2
python-socketio==5.8.0